requirements.txt
langchain
langchain-openai
langchain-mongodb
langgraph==1.0.7
pymongo
kiwipiepy
tqdm
python-dotenv
pandas
beautifulsoup4
requests
lxml
aiohttp
openai
chromadb
pinecone-client
motor
sentence-transformers
transformers
torch
streamlit
